[NO-ISSUE] compatibility with all cert-managers
There's two variant of cert-manager available on the openshift
operatorHub, one for upstream (the community one) and one for downstream
built by Red Hat. Depending on which one the cluster administrator did
chose, the deploy and undeploy script must adapt to it. This is what
this PR brings in to the table. Now the scripts are looking first if the
upstream version is installed, and then if the downstream version is
there. The deploy script will fail if no cert-manager are present on the
cluster.